requirements.txt
Flask==2.0.1
pandas==1.3.3
firebase-admin==5.0.0
openai==0.8.3
python-dotenv==0.19.0
